Keep 64-bit byte counts for the access log
The per-connection byte counters were size_t, so on 32-bit systems a
connection moving more than 4 GiB was logged with its count wrapped,
including the bytes moved by the kernel with SO_SPLICE on OpenBSD.

Make them 64-bit on every system.

+++ src/buffer.h
@@ -28,6 +28,7 @@
 #define BUFFER_H
 
 #include <sys/types.h>
+#include <stdint.h>
 #include <ev.h>
 
 
@@ -40,8 +41,8 @@ struct Buffer {
     size_t max_size;        /* hard cap for buffer_size() */
     ev_tstamp last_recv;
     ev_tstamp last_send;
-    size_t tx_bytes;
-    size_t rx_bytes;
+    uint64_t tx_bytes;      /* wider than size_t on 32-bit systems */
+    uint64_t rx_bytes;
     int pool_managed;
 };
